requirements.txt
antlr4-python3-runtime==4.9.3
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
comet-ml==3.47.3
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
docker-pycreds==0.4.0
dulwich==0.22.6
everett==3.1.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.3
hydra-core==1.3.2
identify==2.6.3
idna==3.10
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lightning-utilities==0.11.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
protobuf==5.28.3
psutil==6.1.0
Pygments==2.18.0
pyparsing==3.2.0
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentry-sdk==2.19.0
setproctitle==1.3.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
wandb==0.18.7
wrapt==1.17.0
wurlitzer==3.1.1
